% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_drought_vhi_risk.R
\name{calc_drought_vhi_risk}
\alias{calc_drought_vhi_risk}
\title{Calculate \% of drought areas (in total areas) or total sum of population using the Vegetation Health Index (https://www.star.nesdis.noaa.gov/smcd/emb/vci/VH/vh_browse.php).}
\usage{
calc_drought_vhi_risk(
  sf,
  start_year,
  end_year,
  threshold,
  ag,
  pop,
  ag_wt = FALSE,
  pop_wt = FALSE
)
}
\arguments{
\item{sf}{object of class sf containing administrative boundaries.}

\item{start_year}{object of class integer; specify the start year for which the Vegetation Health Index is downloaded.}

\item{end_year}{object of class integer; specify the end year for which the Vegetation Health Index is downloaded.}

\item{threshold}{specify a threshold used to define below what value of VHI is considered to be in drought.}

\item{ag}{object of class raster; cropland classification coded 1 if a given pixel is cropland, NA otherwise.}

\item{pop}{object of class raster; population raster to be used so that total number of drought affected people can be computed.}

\item{ag_wt}{TRUE if analysis should be done only on agricultural land, as defined by ag.}

\item{pop_wt}{TRUE if the size of population affected by droughts - instead of \% of land - should be returned.}
}
\value{
\code{calc_drought_vhi_risk()} returns a data.frame object with...
}
\description{
Calculate \% of drought areas (in total areas) or total sum of population using the Vegetation Health Index (https://www.star.nesdis.noaa.gov/smcd/emb/vci/VH/vh_browse.php).
}
\examples{
\dontrun{
download_vhi_files(2000,2021,output_folder)
}
}
