% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fathom_flood_risk2.R
\name{calc_fathom_flood_risk2}
\alias{calc_fathom_flood_risk2}
\title{Calculate flood risks using different metrics based on two different flood raster layers.}
\usage{
calc_fathom_flood_risk2(admin, pop, flood, flood2, depth, resample = FALSE)
}
\arguments{
\item{admin}{object of class sf.}

\item{pop}{object of class raster.}

\item{flood}{object of class raster; flood raster with an expected depth of water for a certain return period.}

\item{flood2}{object of class raster; the second flood raster with an expected depth of water for a certain return period.}

\item{depth}{numeric; depth water depth used as a threshold to define flood.}

\item{resample}{resample flood to match the resolution of pop; default FALSE. If set TRUE, this can take a lot of time depending on the size and resolution of pop.}
}
\value{
\code{calc_flood_risk2()} returns a data.frame object with row ID of admin (id), population (pop), population exposed to flood risks (flood_pop), and percent of people exposed to flood risks (flood_pop_s).
}
\description{
Calculate flood risks using different metrics based on two different flood raster layers.
}
\examples{
\dontrun{
admin <- st_read(paste0("data/moz_adm_20190607b_shp/moz_admbnda_adm1_ine_20190607.shp"))
pop <- raster(paste0("data/MOZ_population_v1_1_gridded.tif"))
flood <- raster::raster(paste0("data/fluvial_undefended/FU_1in20.tif"))
flood2 <- raster::raster(paste0("data/pluvial/P_1in20.tif"))
flood_risk <- calc_flood_risk2(admin, pop, flood, flood2, 0.1)
}
}
