% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_drought_vhi_risk.R
\name{calc_drought_vhi_risk}
\alias{calc_drought_vhi_risk}
\title{Calculate \% of drought areas (in total areas) or total sum of population using the Vegetation Health Index (https://www.star.nesdis.noaa.gov/smcd/emb/vci/VH/vh_TechniqueBackground.php).The VH are indices, which range from 0 to 100 characterizing changes in vegetation conditions from extremely poor (0) to excellent (100). Fair conditions are coded by green color (50), which changes to brown and red when conditions deteriorate and to blue when they improve. The VH values below 40 are used for identifying vegetation stress which is an indirect drought indicator. The VH is very useful for early drought detection, assessing drought area coverage, duration, and intensity, and for monitoring drought impacts on vegetation and agricultural crops.}
\usage{
calc_drought_vhi_risk(
  sf,
  start_year,
  end_year,
  threshold,
  ag,
  pop,
  ag_wt = FALSE,
  pop_wt = FALSE
)
}
\arguments{
\item{sf}{object of class sf containing administrative boundaries.}

\item{start_year}{object of class integer; specify the start year for which the Vegetation Health Index is loaded.}

\item{end_year}{object of class integer; specify the end year for which the Vegetation Health Index is loaded.}

\item{threshold}{object of class numeric; specify a threshold used to define below what value of VHI is considered to be in drought. For more details on how the VHI is used to define drought conditions, see https://www.star.nesdis.noaa.gov/smcd/emb/vci/VH/vh_TechniqueBackground.php.}

\item{ag}{object of class raster; cropland classification coded 1 if a given pixel is cropland, NA otherwise.}

\item{pop}{object of class raster; population raster to be used so that total number of drought affected people can be computed.}

\item{ag_wt}{TRUE if analysis should be done only on agricultural land, as defined by ag.}

\item{pop_wt}{TRUE if the size of population affected by droughts - instead of \% of land - should be returned.}
}
\value{
\code{calc_drought_vhi_risk()} returns a data.frame object with...
}
\description{
Calculate \% of drought areas (in total areas) or total sum of population using the Vegetation Health Index (https://www.star.nesdis.noaa.gov/smcd/emb/vci/VH/vh_TechniqueBackground.php).The VH are indices, which range from 0 to 100 characterizing changes in vegetation conditions from extremely poor (0) to excellent (100). Fair conditions are coded by green color (50), which changes to brown and red when conditions deteriorate and to blue when they improve. The VH values below 40 are used for identifying vegetation stress which is an indirect drought indicator. The VH is very useful for early drought detection, assessing drought area coverage, duration, and intensity, and for monitoring drought impacts on vegetation and agricultural crops.
}
\examples{
\dontrun{
sf <- st_read(paste0("inst/extdata/moz_adm_20190607b_shp/moz_admbnda_adm1_ine_20190607.shp"))
devdisaster::calc_drought_vhi_risk(sf,start_year=2001,end_year=2001,threshold=40)
}
}
